05_chatbot/pyproject.toml
[project]
name = "05-chatbot"
version = "0.1.0"
description = "A chatbot project using OpenAI's API"
readme = "README.md"
authors = [
    { name = "wajihh", email = "wajihh57@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0"
]

[project.scripts]
chatbot = "chatbot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chatbot"]
